+++ src/components/ListingGrid/ListingGridData.jsx
@@ -1,8 +1,0 @@
-import React from 'react';
-import ListingData from '@plone/volto/components/manage/Blocks/Listing/ListingData';
-
-const ListingGridData = (props) => {
-  return <ListingData {...props} />;
-};
-
-export default ListingGridData;